% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRollingWindowAKI.R
\name{addRollingWindowAKI}
\alias{addRollingWindowAKI}
\title{Rolling Window AKI}
\usage{
addRollingWindowAKI(
  dataframe,
  window1 = as.difftime(2, units = "days"),
  window2 = as.difftime(7, units = "days")
)
}
\arguments{
\item{dataframe}{The patient dataset}

\item{window1}{The rolling window length of the shorter time window; defaults to 48 hours}

\item{window2}{The rolling window length of the longer time window; defaults to 162 hours}
}
\value{
The patient dataset with the rolling-window AKI column added in
}
\description{
Rolling Window AKI
}
\examples{
\dontrun{
addRollingWindowAKI(df)
}
}
