% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnAKIpatients.R
\name{runAllDefinitions}
\alias{runAllDefinitions}
\title{Run all THREE definitions for acute kidney injury (AKI); helper function.}
\usage{
runAllDefinitions(dataframe, padding = c(4, "hours"))
}
\arguments{
\item{dataframe}{Your input dataset, of class c(data.table, data.frame).}

\item{padding}{The amount of padding you would like to add to the rolling windows. Enter this as a c(\code{integer}, \code{string}) vector where
\code{integer} is the amount of time and \code{string} are the units of time. Defaults to \code{c(4L, "hours")}.}
}
\value{
The input patient dataset with the 3 AKI columns added in, titled \code{RMW}, \code{HBT} and \code{BCI}.
}
\description{
Run all THREE definitions for acute kidney injury (AKI); helper function.
}
\examples{
runAllDefinitions(toy)
}
